hub/ai-services/vision/requirements.txt
# AI Vision Service Dependencies
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0

# Computer Vision - Open Source Stack
ultralytics==8.0.206  # YOLOv8
opencv-python==4.8.1.78
mediapipe==0.10.7
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0

# Image processing
scikit-image==0.22.0
albumentations==1.3.1

# HTTP client for API calls
httpx==0.25.0
requests==2.31.0
aiofiles==23.2.1

# System monitoring
psutil==5.9.6
aiofiles==23.2.1

# Configuration and utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1